refactor(sdk-logs): undo validation file rename
