+++ experimental/packages/sdk-logs/src/utils/validation.ts
